--- a/Cargo.toml
+++ b/Cargo.toml
@@ -42,6 +42,7 @@ rust_decimal = { version = "1.26.1", optional = true }
 regex = { version = "1.6.0", optional = true }
 time = { version = "0.3.14", features = ["formatting"], optional = true }
 paste = { version = "1.0.9", optional = true }
+bitvec = "1.0.1"
 
 # TODO: Remove these dependencies
 rand = { version = "0.8", optional = true }