--- a/benches/consolidation.rs
+++ b/benches/consolidation.rs
@@ -1,3 +1,4 @@
+use bitvec::vec::BitVec;
 use criterion::{black_box, criterion_group, criterion_main, BatchSize, Criterion};
 use dbsp::{
     algebra::{AddAssignByRef, HasZero},
@@ -160,6 +161,52 @@ macro_rules! consolidation_benches {
                     b.iter_batched(
                         || unsorted.clone(),
                         |mut unsorted| black_box(&mut unsorted).sort_by(|(key1, _), (key2, _)| key1.cmp(key2)),
+                        BatchSize::PerIteration,
+                    );
+                });
+            )*
+            group.finish();
+
+            let mut group = c.benchmark_group("shuffle-by-indices");
+            group.sample_size(10);
+            $(
+                group.bench_function($name, |b| {
+                    let mut rng = Xoshiro256StarStar::from_seed(SEED);
+
+                    let indices: Vec<usize> = (0..$size).collect();
+                    let mut keys: Vec<usize> = (0..$size).map(|_| rng.gen()).collect();
+                    keys.shuffle(&mut rng);
+                    let mut diffs: Vec<isize> = (0..$size).map(|_| rng.gen()).collect();
+                    diffs.shuffle(&mut rng);
+
+                    b.iter_batched(
+                        || (keys.clone(), diffs.clone(), indices.clone()),
+                        |(mut keys, mut diffs, mut indices)| {
+                            unsafe { consolidation::shuffle_by_indices(&mut keys, &mut diffs, &mut indices) };
+                        },
+                        BatchSize::PerIteration,
+                    );
+                });
+            )*
+            group.finish();
+
+            let mut group = c.benchmark_group("shuffle-by-indices-bitvec");
+            group.sample_size(10);
+            $(
+                group.bench_function($name, |b| {
+                    let mut rng = Xoshiro256StarStar::from_seed(SEED);
+
+                    let indices: Vec<usize> = (0..$size).collect();
+                    let mut keys: Vec<usize> = (0..$size).map(|_| rng.gen()).collect();
+                    keys.shuffle(&mut rng);
+                    let mut diffs: Vec<isize> = (0..$size).map(|_| rng.gen()).collect();
+                    diffs.shuffle(&mut rng);
+
+                    b.iter_batched(
+                        || (keys.clone(), diffs.clone(), indices.clone(), BitVec::repeat(false, $size).into_boxed_bitslice()),
+                        |(mut keys, mut diffs, mut indices, mut visited)| {
+                            unsafe { consolidation::shuffle_by_indices_bitvec(&mut keys, &mut diffs, &mut indices, &mut visited) };
+                        },
                         BatchSize::PerIteration,
                     );
                 });