--- a/benches/consolidation.rs
+++ b/benches/consolidation.rs
@@ -173,7 +173,8 @@ macro_rules! consolidation_benches {
                 group.bench_function($name, |b| {
                     let mut rng = Xoshiro256StarStar::from_seed(SEED);
 
-                    let indices: Vec<usize> = (0..$size).collect();
+                    let mut indices: Vec<usize> = (0..$size).collect();
+                    indices.shuffle(&mut rng);
                     let mut keys: Vec<usize> = (0..$size).map(|_| rng.gen()).collect();
                     keys.shuffle(&mut rng);
                     let mut diffs: Vec<isize> = (0..$size).map(|_| rng.gen()).collect();
@@ -196,7 +197,8 @@ macro_rules! consolidation_benches {
                 group.bench_function($name, |b| {
                     let mut rng = Xoshiro256StarStar::from_seed(SEED);
 
-                    let indices: Vec<usize> = (0..$size).collect();
+                    let mut indices: Vec<usize> = (0..$size).collect();
+                    indices.shuffle(&mut rng);
                     let mut keys: Vec<usize> = (0..$size).map(|_| rng.gen()).collect();
                     keys.shuffle(&mut rng);
                     let mut diffs: Vec<isize> = (0..$size).map(|_| rng.gen()).collect();
@@ -206,6 +208,30 @@ macro_rules! consolidation_benches {
                         || (keys.clone(), diffs.clone(), indices.clone(), BitVec::repeat(false, $size).into_boxed_bitslice()),
                         |(mut keys, mut diffs, mut indices, mut visited)| {
                             unsafe { consolidation::utils::shuffle_by_indices_bitvec(&mut keys, &mut diffs, &mut indices, &mut visited) };
+                        },
+                        BatchSize::PerIteration,
+                    );
+                });
+            )*
+            group.finish();
+
+            let mut group = c.benchmark_group("consolidate-paired-vecs");
+            group.sample_size(10);
+            $(
+                group.bench_function($name, |b| {
+                    let mut rng = Xoshiro256StarStar::from_seed(SEED);
+
+                    let mut indices: Vec<usize> = (0..$size).collect();
+                    indices.shuffle(&mut rng);
+                    let mut keys: Vec<usize> = (0..$size).map(|_| rng.gen()).collect();
+                    keys.shuffle(&mut rng);
+                    let mut diffs: Vec<isize> = (0..$size).map(|_| rng.gen()).collect();
+                    diffs.shuffle(&mut rng);
+
+                    b.iter_batched(
+                        || (keys.clone(), diffs.clone(), indices.clone()),
+                        |(mut keys, mut diffs, mut indices)| {
+                            consolidation::consolidate_paired_vecs_from(&mut keys, &mut diffs, &mut indices, 0);
                         },
                         BatchSize::PerIteration,
                     );