@Kixiron 2022-09-21
Got perf numbers from the current master, this PR introduces almost unilateral perf gains apart from the cost of pushing small (<100k) numbers of elements which is to be expected since vector growth has a much larger impact on smaller cases until it gets amortized away

```
add/0
time:   [37.659 ns 37.943 ns 38.092 ns]

add/10
time:   [255.41 ns 257.54 ns 260.40 ns]

add/100
time:   [858.30 ns 860.81 ns 863.43 ns]

add/1000
time:   [7.9732 µs 8.0161 µs 8.0566 µs]

add/10,000
time:   [83.199 µs 83.824 µs 84.269 µs]

add/100,000
time:   [874.96 µs 879.34 µs 884.38 µs]

add/1,000,000
time:   [11.627 ms 11.837 ms 12.131 ms]

add/10,000,000
time:   [104.80 ms 105.58 ms 106.68 ms]

add/100,000,000
time:   [1.0393 s 1.0521 s 1.0676 s]
```

```
add-by-ref/0
time:   [45.870 ns 45.941 ns 45.992 ns]

add-by-ref/10
time:   [166.67 ns 167.35 ns 167.71 ns]

add-by-ref/100
time:   [765.37 ns 770.40 ns 775.25 ns]

add-by-ref/1000
time:   [7.7521 µs 7.7743 µs 7.8039 µs]

add-by-ref/10,000
time:   [81.201 µs 81.396 µs 81.561 µs]

add-by-ref/100,000
time:   [824.76 µs 829.05 µs 832.64 µs]

add-by-ref/1,000,000
time:   [9.5393 ms 9.5919 ms 9.6416 ms]

add-by-ref/10,000,000
time:   [93.050 ms 93.826 ms 95.101 ms]

add-by-ref/100,000,000
time:   [958.83 ms 970.60 ms 987.32 ms]
```

```
add-assign/0
time:   [37.044 ns 37.197 ns 37.354 ns]

add-assign/10
time:   [268.54 ns 271.34 ns 274.43 ns]

add-assign/100
time:   [884.43 ns 890.04 ns 894.91 ns]

add-assign/1000
time:   [8.0023 µs 8.0406 µs 8.0906 µs]

add-assign/10,000
time:   [83.438 µs 84.199 µs 84.828 µs]

add-assign/100,000
time:   [832.34 µs 834.47 µs 838.66 µs]

add-assign/1,000,000
time:   [10.859 ms 10.898 ms 10.930 ms]

add-assign/10,000,000
time:   [105.32 ms 105.95 ms 106.30 ms]

add-assign/100,000,000
time:   [1.1678 s 1.2033 s 1.2482 s]
```

```
add-assign-by-ref/0
time:   [36.253 ns 36.306 ns 36.369 ns]

add-assign-by-ref/10
time:   [243.65 ns 249.73 ns 254.79 ns]

add-assign-by-ref/100
time:   [850.63 ns 855.76 ns 860.06 ns]

add-assign-by-ref/1000
time:   [7.9372 µs 7.9703 µs 8.0069 µs]

add-assign-by-ref/10,000
time:   [82.812 µs 83.071 µs 83.362 µs]

add-assign-by-ref/100,000
time:   [822.95 µs 827.49 µs 832.29 µs]

add-assign-by-ref/1,000,000
time:   [10.726 ms 10.796 ms 10.843 ms]

add-assign-by-ref/10,000,000
time:   [104.16 ms 105.33 ms 106.38 ms]

add-assign-by-ref/100,000,000
time:   [1.1058 s 1.1187 s 1.1316 s]
```

```
neg/0
time:   [35.936 ns 36.042 ns 36.142 ns]

neg/10
time:   [37.274 ns 37.386 ns 37.464 ns]

neg/100
time:   [46.009 ns 46.082 ns 46.174 ns]

neg/1000
time:   [171.26 ns 171.50 ns 171.77 ns]

neg/10,000
time:   [1.4330 µs 1.4356 µs 1.4397 µs]

neg/100,000
time:   [14.805 µs 14.899 µs 15.051 µs]

neg/1,000,000
time:   [358.93 µs 363.07 µs 370.51 µs]

neg/10,000,000
time:   [4.9827 ms 5.0703 ms 5.1756 ms]

neg/100,000,000
time:   [55.118 ms 56.504 ms 58.263 ms]
```

```
neg-by-ref/0
time:   [45.223 ns 45.260 ns 45.311 ns]

neg-by-ref/10
time:   [98.202 ns 98.909 ns 99.816 ns]

neg-by-ref/100
time:   [116.58 ns 118.19 ns 119.82 ns]

neg-by-ref/1000
time:   [376.14 ns 391.80 ns 407.49 ns]

neg-by-ref/10,000
time:   [3.8534 µs 3.8741 µs 3.8935 µs]

neg-by-ref/100,000
time:   [36.027 µs 36.319 µs 36.658 µs]

neg-by-ref/1,000,000
time:   [2.1129 ms 2.1295 ms 2.1409 ms]

neg-by-ref/10,000,000
time:   [22.470 ms 22.636 ms 22.792 ms]

neg-by-ref/100,000,000
time:   [245.92 ms 249.27 ms 253.47 ms]
```

```
build-unordered-pushing/0
time:   [35.889 ns 35.951 ns 36.014 ns]

build-unordered-pushing/10
time:   [272.04 ns 273.64 ns 275.20 ns]

build-unordered-pushing/100
time:   [601.11 ns 602.92 ns 605.23 ns]

build-unordered-pushing/1000
time:   [2.5626 µs 2.5668 µs 2.5712 µs]

build-unordered-pushing/10,000
time:   [22.226 µs 22.901 µs 23.525 µs]

build-unordered-pushing/100,000
time:   [594.64 µs 603.66 µs 614.46 µs]

build-unordered-pushing/1,000,000
time:   [6.9153 ms 6.9397 ms 6.9642 ms]

build-unordered-pushing/10,000,000
time:   [89.800 ms 90.445 ms 91.134 ms]

build-unordered-pushing/100,000,000
time:   [1.0232 s 1.0371 s 1.0514 s]
```

```
build-unordered-boundary/0
time:   [38.582 ns 38.697 ns 38.769 ns]

build-unordered-boundary/10
time:   [92.162 ns 92.424 ns 92.834 ns]

build-unordered-boundary/100
time:   [905.29 ns 917.84 ns 927.16 ns]

build-unordered-boundary/1000
time:   [11.374 µs 11.413 µs 11.474 µs]

build-unordered-boundary/10,000
time:   [225.32 µs 226.72 µs 227.52 µs]

build-unordered-boundary/100,000
time:   [2.7121 ms 2.7674 ms 2.8007 ms]

build-unordered-boundary/1,000,000
time:   [33.926 ms 34.245 ms 34.503 ms]

build-unordered-boundary/10,000,000
time:   [387.22 ms 391.93 ms 396.67 ms]

build-unordered-boundary/100,000,000
time:   [4.5041 s 4.5539 s 4.6049 s]
```

```
build-unordered-done/0
time:   [51.940 ns 52.044 ns 52.137 ns]

build-unordered-done/10
time:   [187.17 ns 187.98 ns 189.40 ns]

build-unordered-done/100
time:   [1.1537 µs 1.1577 µs 1.1647 µs]

build-unordered-done/1000
time:   [11.765 µs 11.812 µs 11.843 µs]

build-unordered-done/10,000
time:   [234.90 µs 235.76 µs 236.69 µs]

build-unordered-done/100,000
time:   [2.9112 ms 2.9337 ms 2.9624 ms]

build-unordered-done/1,000,000
time:   [36.135 ms 36.391 ms 36.547 ms]

build-unordered-done/10,000,000
time:   [428.33 ms 434.04 ms 439.59 ms]

build-unordered-done/100,000,000
time:   [4.8069 s 4.8641 s 4.9200 s]
```

```
ordered-builder-push-merge/0
time:   [39.689 ns 39.760 ns 39.842 ns]

ordered-builder-push-merge/10
time:   [232.10 ns 232.62 ns 233.22 ns]

ordered-builder-push-merge/100
time:   [870.33 ns 871.52 ns 872.90 ns]

ordered-builder-push-merge/1000
time:   [8.2505 µs 8.2661 µs 8.2837 µs]

ordered-builder-push-merge/10,000
time:   [86.577 µs 86.764 µs 86.997 µs]

ordered-builder-push-merge/100,000
time:   [862.92 µs 864.43 µs 865.81 µs]

ordered-builder-push-merge/1,000,000
time:   [10.682 ms 10.711 ms 10.742 ms]

ordered-builder-push-merge/10,000,000
time:   [105.69 ms 106.14 ms 106.63 ms]

ordered-builder-push-merge/100,000,000
time:   [1.1132 s 1.1171 s 1.1208 s]
```