@Kixiron 2022-09-26
Current nexmark benches show little to no change for me

Main:
```
┌───────┬─────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┐
│ Query │ #Events     │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Max RSS(Kb) │
├───────┼─────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┤
│ q0    │ 100,000,000 │ 8     │ 14.492s │ 115.933s        │ 862.564 K/s      │ 125.390s      │ 0.375s        │ 127,620     │
│ q1    │ 100,000,000 │ 8     │ 14.516s │ 116.127s        │ 861.129 K/s      │ 139.375s      │ 0.188s        │ N/A         │
│ q2    │ 100,000,000 │ 8     │ 14.410s │ 115.280s        │ 867.452 K/s      │ 128.078s      │ 0.359s        │ N/A         │
│ q3    │ 100,000,000 │ 8     │ 14.713s │ 117.702s        │ 849.606 K/s      │ 130.782s      │ 0.406s        │ N/A         │
│ q4    │ 100,000,000 │ 8     │ 16.932s │ 135.458s        │ 738.236 K/s      │ 166.953s      │ 4.953s        │ N/A         │
│ q5    │ 100,000,000 │ 8     │ 14.407s │ 115.253s        │ 867.654 K/s      │ 133.078s      │ 0.844s        │ N/A         │
│ q6    │ 100,000,000 │ 8     │ 16.727s │ 133.813s        │ 747.311 K/s      │ 172.969s      │ 3.000s        │ N/A         │
│ q7    │ 100,000,000 │ 8     │ 20.739s │ 165.913s        │ 602.727 K/s      │ 205.328s      │ 5.313s        │ N/A         │
│ q8    │ 100,000,000 │ 8     │ 15.222s │ 121.780s        │ 821.153 K/s      │ 132.781s      │ 0.547s        │ N/A         │
│ q9    │ 100,000,000 │ 8     │ 74.303s │ 594.420s        │ 168.231 K/s      │ 346.422s      │ 5.296s        │ N/A         │
│ q12   │ 100,000,000 │ 8     │ 15.296s │ 122.367s        │ 817.211 K/s      │ 143.422s      │ 0.469s        │ N/A         │
│ q13   │ 100,000,000 │ 8     │ 21.014s │ 168.113s        │ 594.839 K/s      │ 206.468s      │ 3.344s        │ N/A         │
│ q14   │ 100,000,000 │ 8     │ 14.464s │ 115.708s        │ 864.242 K/s      │ 135.329s      │ 0.156s        │ N/A         │
│ q15   │ 100,000,000 │ 8     │ 15.275s │ 122.202s        │ 818.318 K/s      │ 167.437s      │ 2.063s        │ N/A         │
│ q16   │ 100,000,000 │ 8     │ 50.934s │ 407.470s        │ 245.417 K/s      │ 402.016s      │ 5.234s        │ N/A         │
│ q17   │ 100,000,000 │ 8     │ 21.841s │ 174.731s        │ 572.310 K/s      │ 247.672s      │ 3.766s        │ N/A         │
│ q18   │ 100,000,000 │ 8     │ 44.364s │ 354.912s        │ 281.760 K/s      │ 240.656s      │ 4.781s        │ N/A         │
│ q19   │ 100,000,000 │ 8     │ 44.184s │ 353.474s        │ 282.906 K/s      │ 327.969s      │ 12.328s       │ N/A         │
│ q20   │ 100,000,000 │ 8     │ 32.026s │ 256.206s        │ 390.311 K/s      │ 232.265s      │ 4.531s        │ N/A         │
│ q21   │ 100,000,000 │ 8     │ 15.673s │ 125.385s        │ 797.545 K/s      │ 143.610s      │ 0.344s        │ N/A         │
│ q22   │ 100,000,000 │ 8     │ 14.910s │ 119.281s        │ 838.358 K/s      │ 144.328s      │ 0.203s        │ N/A         │
└───────┴─────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┘
```

This Branch:
```
┌───────┬─────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬─────────────┐
│ Query │ #Events     │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Max RSS(Kb) │
├───────┼─────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼─────────────┤
│ q0    │ 100,000,000 │ 8     │ 14.247s │ 113.980s        │ 877.349 K/s      │ 123.313s      │ 0.296s        │ 148,196     │
│ q1    │ 100,000,000 │ 8     │ 14.618s │ 116.941s        │ 855.129 K/s      │ 136.812s      │ 0.266s        │ N/A         │
│ q2    │ 100,000,000 │ 8     │ 15.008s │ 120.064s        │ 832.888 K/s      │ 127.828s      │ 0.328s        │ N/A         │
│ q3    │ 100,000,000 │ 8     │ 14.452s │ 115.613s        │ 864.955 K/s      │ 128.703s      │ 0.625s        │ N/A         │
│ q4    │ 100,000,000 │ 8     │ 16.834s │ 134.669s        │ 742.563 K/s      │ 168.219s      │ 4.922s        │ N/A         │
│ q5    │ 100,000,000 │ 8     │ 14.368s │ 114.943s        │ 869.998 K/s      │ 128.563s      │ 0.594s        │ N/A         │
│ q6    │ 100,000,000 │ 8     │ 16.808s │ 134.460s        │ 743.715 K/s      │ 179.203s      │ 3.265s        │ N/A         │
│ q7    │ 100,000,000 │ 8     │ 21.374s │ 170.990s        │ 584.829 K/s      │ 209.125s      │ 5.844s        │ N/A         │
│ q8    │ 100,000,000 │ 8     │ 15.061s │ 120.487s        │ 829.965 K/s      │ 132.547s      │ 0.360s        │ N/A         │
│ q9    │ 100,000,000 │ 8     │ 75.226s │ 601.810s        │ 166.165 K/s      │ 349.140s      │ 6.421s        │ N/A         │
│ q12   │ 100,000,000 │ 8     │ 15.426s │ 123.410s        │ 810.307 K/s      │ 141.282s      │ 0.454s        │ N/A         │
│ q13   │ 100,000,000 │ 8     │ 21.489s │ 171.915s        │ 581.683 K/s      │ 208.093s      │ 3.703s        │ N/A         │
│ q14   │ 100,000,000 │ 8     │ 14.723s │ 117.784s        │ 849.013 K/s      │ 133.750s      │ 0.265s        │ N/A         │
│ q15   │ 100,000,000 │ 8     │ 15.207s │ 121.659s        │ 821.972 K/s      │ 167.407s      │ 1.969s        │ N/A         │
│ q16   │ 100,000,000 │ 8     │ 50.642s │ 405.138s        │ 246.830 K/s      │ 410.265s      │ 4.000s        │ N/A         │
│ q17   │ 100,000,000 │ 8     │ 21.890s │ 175.116s        │ 571.048 K/s      │ 253.188s      │ 3.500s        │ N/A         │
│ q18   │ 100,000,000 │ 8     │ 43.835s │ 350.682s        │ 285.159 K/s      │ 245.703s      │ 5.719s        │ N/A         │
│ q19   │ 100,000,000 │ 8     │ 43.663s │ 349.301s        │ 286.286 K/s      │ 325.437s      │ 10.000s       │ N/A         │
│ q20   │ 100,000,000 │ 8     │ 32.045s │ 256.361s        │ 390.076 K/s      │ 231.469s      │ 4.047s        │ N/A         │
│ q21   │ 100,000,000 │ 8     │ 15.759s │ 126.076s        │ 793.174 K/s      │ 139.672s      │ 0.375s        │ N/A         │
│ q22   │ 100,000,000 │ 8     │ 14.044s │ 112.355s        │ 890.036 K/s      │ 132.281s      │ 0.312s        │ N/A         │
└───────┴─────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴─────────────┘
```